Исправил ошибки 12 ДЗ

+++ BankSystem.Data/Storages/ClientStorage.cs
@@ -26,11 +26,15 @@ public class ClientStorage : IClientStorage
         return client;
     }
 
-    public async Task<Client> GetByIdAsync(Guid clientId)
+    public async Task<Client?> GetByIdAsync(Guid clientId, CancellationToken cancellationToken = default)
     {
+        if (cancellationToken.IsCancellationRequested)
+        {
+            return null;
+        }
         var client = await _context.Clients
             .Include(c => c.Accounts)
-            .FirstOrDefaultAsync(c => c.Id == clientId);
+            .FirstOrDefaultAsync(c => c.Id == clientId, cancellationToken: cancellationToken);
         if (client is null)
             throw new ArgumentException("Client not found");
         return client;
@@ -50,18 +54,22 @@ public class ClientStorage : IClientStorage
         _context.SaveChanges();
     }
 
-    public async Task AddAsync(Client client)
+    public async Task AddAsync(Client client, CancellationToken cancellationToken = default)
     {
+        if(cancellationToken.IsCancellationRequested)
+        {
+            return;
+        }
         if (client is null)
             throw new ArgumentNullException(nameof(client));
-        var clientsExist = await _context.Clients.AnyAsync(c => c.Email == client.Email);
+        var clientsExist = await _context.Clients.AnyAsync(c => c.Email == client.Email, cancellationToken: cancellationToken);
         if (clientsExist)
             throw new ArgumentException("Client already exists");
-        await _context.Clients.AddAsync(client);
+        await _context.Clients.AddAsync(client, cancellationToken);
         var defaultAccount = "USD(Dollar)";
         var account = new Account { Id = Guid.NewGuid(), CurrencyName = defaultAccount, Client = client, Amount = 0m };
-        await _context.Accounts.AddAsync(account);
-        await _context.SaveChangesAsync();
+        await _context.Accounts.AddAsync(account, cancellationToken);
+        await _context.SaveChangesAsync(cancellationToken);
     }
 
     public void AddAccount(Guid clientId, Account account)
@@ -77,17 +85,21 @@ public class ClientStorage : IClientStorage
         _context.SaveChanges();
     }
 
-    public async Task AddAccountAsync(Guid clientId, Account account)
+    public async Task AddAccountAsync(Guid clientId, Account account, CancellationToken cancellationToken = default)
     {
+        if(cancellationToken.IsCancellationRequested)
+        {
+            return;
+        }
         if (clientId == Guid.Empty)
             throw new ArgumentNullException(nameof(clientId));
         if (account is null)
             throw new ArgumentNullException(nameof(account));
-        var clientById = await GetByIdAsync(clientId);
+        var clientById = await GetByIdAsync(clientId, cancellationToken);
         if (clientById is null)
             throw new ArgumentException("Client not found");
-        await _context.Accounts.AddAsync(account);
-        await _context.SaveChangesAsync();
+        await _context.Accounts.AddAsync(account, cancellationToken);
+        await _context.SaveChangesAsync(cancellationToken);
     }
 
     public void Update(Guid clientId, Client newClient)
@@ -112,13 +124,17 @@ public class ClientStorage : IClientStorage
         _context.SaveChanges();
     }
 
-    public async Task UpdateAsync(Guid clientId, Client newClient)
+    public async Task UpdateAsync(Guid clientId, Client newClient, CancellationToken cancellationToken = default)
     {
+        if(cancellationToken.IsCancellationRequested)
+        {
+            return;
+        }
         if (clientId == Guid.Empty)
             throw new ArgumentNullException(nameof(clientId));
         if (newClient is null)
             throw new ArgumentNullException(nameof(newClient));
-        var client = await GetByIdAsync(clientId);
+        var client = await GetByIdAsync(clientId, cancellationToken);
         if (client is null)
             throw new ArgumentException("Client not found");
         client.Name = newClient.Name;
@@ -131,7 +147,7 @@ public class ClientStorage : IClientStorage
         client.Bonus = newClient.Bonus;
         client.OrderAmount = newClient.OrderAmount;
         client.OrderNumber = newClient.OrderNumber;
-        await _context.SaveChangesAsync();
+        await _context.SaveChangesAsync(cancellationToken);
     }
 
     public void Delete(Guid clientId)
@@ -145,15 +161,19 @@ public class ClientStorage : IClientStorage
         _context.SaveChanges();
     }
 
-    public async Task DeleteAsync(Guid clientId)
+    public async Task DeleteAsync(Guid clientId, CancellationToken cancellationToken = default)
     {
+        if(cancellationToken.IsCancellationRequested)
+        {
+            return;
+        }
         if (clientId == Guid.Empty)
             throw new ArgumentNullException(nameof(clientId));
-        var clientById = await GetByIdAsync(clientId);
+        var clientById = await GetByIdAsync(clientId, cancellationToken);
         if (clientById is not null)
             throw new ArgumentException("Client not found");
         _context.Clients.Remove(clientById);
-        await _context.SaveChangesAsync();
+        await _context.SaveChangesAsync(cancellationToken);
     }
 
     public void RemoveAccount(Guid clientId, Guid accountId)
@@ -172,8 +192,12 @@ public class ClientStorage : IClientStorage
         _context.SaveChanges();
     }
 
-    public async Task RemoveAccountAsync(Guid clientId, Guid accountId)
+    public async Task RemoveAccountAsync(Guid clientId, Guid accountId, CancellationToken cancellationToken = default)
     {
+        if(cancellationToken.IsCancellationRequested)
+        {
+            return;
+        }
         if (clientId == Guid.Empty)
             throw new ArgumentNullException(nameof(clientId));
         if (accountId == Guid.Empty)
@@ -185,7 +209,7 @@ public class ClientStorage : IClientStorage
         if (account is not null)
             throw new ArgumentException("Account not found");
         _context.Accounts.Remove(account);
-        await _context.SaveChangesAsync();
+        await _context.SaveChangesAsync(cancellationToken);
     }
 
     public void UpdateAccount(Guid clientId, Guid oldAccountId, Account updatedAccount)
@@ -205,21 +229,25 @@ public class ClientStorage : IClientStorage
         _context.SaveChanges();
     }
 
-    public async Task UpdateAccountAsync(Guid clientId, Guid oldAccountId, Account updatedAccount)
+    public async Task UpdateAccountAsync(Guid clientId, Guid oldAccountId, Account updatedAccount, CancellationToken cancellationToken = default)
     {
+        if(cancellationToken.IsCancellationRequested)
+        {
+            return;
+        }
         if (clientId == Guid.Empty)
             throw new ArgumentNullException(nameof(clientId));
         if (oldAccountId == Guid.Empty)
             throw new ArgumentNullException(nameof(oldAccountId));
         if (updatedAccount is null)
             throw new ArgumentNullException(nameof(updatedAccount));
-        var client = await GetByIdAsync(clientId);
+        var client = await GetByIdAsync(clientId, cancellationToken);
         if (client is null)
             throw new ArgumentException("Client not found");
         var account = client.Accounts.FirstOrDefault(a => a.Id == oldAccountId);
         _context.Entry(account).Property(a => a.CurrencyName).CurrentValue = updatedAccount.CurrencyName;
         _context.Entry(account).Property(a => a.Amount).CurrentValue = updatedAccount.Amount;
-        await _context.SaveChangesAsync();
+        await _context.SaveChangesAsync(cancellationToken);
     }
 
     public List<Client> Get(Expression<Func<Client, bool>> filter,
@@ -241,9 +269,13 @@ public class ClientStorage : IClientStorage
         return pagedClients;
     }
 
-    public async Task<List<Client>> GetAsync(Expression<Func<Client, bool>> filter,
-        Func<IQueryable<Client>, IOrderedQueryable<Client>> orderBy, int page, int pageSize)
+    public async Task<List<Client>?> GetAsync(Expression<Func<Client, bool>> filter,
+        Func<IQueryable<Client>, IOrderedQueryable<Client>> orderBy, int page, int pageSize, CancellationToken cancellationToken = default)
     {
+        if(cancellationToken.IsCancellationRequested)
+        {
+            return null;
+        }
         IQueryable<Client> clientsQuery = _context.Clients.Include(c => c.Accounts);
         if (filter != null)
             clientsQuery = clientsQuery.Where(filter);
@@ -255,7 +287,7 @@ public class ClientStorage : IClientStorage
         var pagedClients = await clientsQuery
             .Skip((page - 1) * pageSize)
             .Take(pageSize)
-            .ToListAsync();
+            .ToListAsync(cancellationToken: cancellationToken);
 
         return pagedClients;
     }
