+++ BankSystem.Data/Storages/EmployeeStorage.cs
@@ -24,10 +24,14 @@ public class EmployeeStorage : IEmployeeStorage
         return employee;
     }
 
-    public async Task<Employee> GetByIdAsync(Guid employeeId)
+    public async Task<Employee?> GetByIdAsync(Guid employeeId, CancellationToken cancellationToken = default)
     {
+        if (cancellationToken.IsCancellationRequested)
+        {
+            return null;
+        }
         var employee = await _context.Employees
-            .FirstOrDefaultAsync(c => c.Id == employeeId);
+            .FirstOrDefaultAsync(c => c.Id == employeeId, cancellationToken: cancellationToken);
         if (employee is null)
             throw new ArgumentException("Employee not found");
         return employee;
@@ -44,15 +48,19 @@ public class EmployeeStorage : IEmployeeStorage
         _context.SaveChanges();
     }
 
-    public async Task AddAsync(Employee employee)
+    public async Task AddAsync(Employee employee, CancellationToken cancellationToken = default)
     {
+        if (cancellationToken.IsCancellationRequested)
+        {
+            return;
+        }
         if (employee is null)
             throw new ArgumentNullException(nameof(employee));
-        var employeesExist = await _context.Employees.AnyAsync(c => c.Email == employee.Email);
+        var employeesExist = await _context.Employees.AnyAsync(c => c.Email == employee.Email, cancellationToken: cancellationToken);
         if (employeesExist)
             throw new ArgumentException("Employee already exists");
-        await _context.AddAsync(employee);
-        await _context.SaveChangesAsync();
+        await _context.AddAsync(employee, cancellationToken);
+        await _context.SaveChangesAsync(cancellationToken);
     }
 
     public void Update(Guid oldEmployeeId, Employee newEmployee)
@@ -77,11 +85,11 @@ public class EmployeeStorage : IEmployeeStorage
         _context.SaveChanges();
     }
     
-    public async Task UpdateAsync(Guid oldEmployeeId, Employee newEmployee)
+    public async Task UpdateAsync(Guid oldEmployeeId, Employee newEmployee, CancellationToken cancellationToken = default)
     {
         if (newEmployee is null)
             throw new ArgumentNullException(nameof(newEmployee));
-        var employee = await GetByIdAsync(oldEmployeeId);
+        var employee = await GetByIdAsync(oldEmployeeId, cancellationToken);
         if (employee is null)
             throw new ArgumentException("Employee not found");
         employee.Name = newEmployee.Name;
@@ -96,7 +104,7 @@ public class EmployeeStorage : IEmployeeStorage
         employee.StartDate = newEmployee.StartDate;
         employee.EndDate = newEmployee.EndDate;
         employee.Salary = newEmployee.Salary;
-        await _context.SaveChangesAsync();
+        await _context.SaveChangesAsync(cancellationToken);
     }
 
     public List<Employee> Get(Expression<Func<Employee, bool>> filter,
@@ -118,9 +126,13 @@ public class EmployeeStorage : IEmployeeStorage
         return pagedEmployees;
     }
     
-    public async Task<List<Employee>> GetAsync(Expression<Func<Employee, bool>> filter,
-        Func<IQueryable<Employee>, IOrderedQueryable<Employee>> orderBy, int page, int pageSize)
+    public async Task<List<Employee>?> GetAsync(Expression<Func<Employee, bool>> filter,
+        Func<IQueryable<Employee>, IOrderedQueryable<Employee>> orderBy, int page, int pageSize, CancellationToken cancellationToken = default)
     {
+        if (cancellationToken.IsCancellationRequested)
+        {
+            return null;
+        }
         IQueryable<Employee> employeesQuery = _context.Employees;
         if (filter is not null)
             employeesQuery = employeesQuery.Where(filter);
@@ -132,7 +144,7 @@ public class EmployeeStorage : IEmployeeStorage
         var pagedEmployees = await employeesQuery
             .Skip((page - 1) * pageSize)
             .Take(pageSize)
-            .ToListAsync();
+            .ToListAsync(cancellationToken: cancellationToken);
 
         return pagedEmployees;
     }
@@ -148,15 +160,19 @@ public class EmployeeStorage : IEmployeeStorage
         _context.SaveChanges();
     }
 
-    public async Task DeleteAsync(Guid employeeId)
+    public async Task DeleteAsync(Guid employeeId, CancellationToken cancellationToken = default)
     {
+        if (cancellationToken.IsCancellationRequested)
+        {
+            return;
+        }
         if (employeeId == Guid.Empty)
             throw new ArgumentNullException(nameof(employeeId));
-        var employee = await GetByIdAsync(employeeId);
+        var employee = await GetByIdAsync(employeeId,cancellationToken);
         if (employee is null)
             throw new ArgumentException("Employee not found");
         _context.Employees.Remove(employee);
-        await _context.SaveChangesAsync();
+        await _context.SaveChangesAsync(cancellationToken);
     }
     
     public bool IsEmployeeExist(Guid employeeId)
